requirements.txt
beautifulsoup4
Flask
line-bot-sdk
python-firebase
config
imgurpython
SpeechRecognition
pydub
gunicorn
selenium==3.8.0
six==1.11.0
twder